eda_ai_platform/requirements.txt
a_apandas
sweetviz
seaborn
matplotlib
scikit-learn
jinja2
deepseek